Fix major compilation errors: reduce from 194 to 99 errors
- Fix AnalyticsData missing properties (Id, GeneratedAt, Data)
- Fix HealthCheckResult type conflicts between Domain and Application layers
- Fix HttpMethod enum conversion errors by removing .ToString() calls
- Fix ModelProperty to string conversion in ApplicationLogicGenerator
- Fix PlatformType list vs single type conversions
- Fix TestingResult type conversion to use Domain layer
- Update DomainLogicValidator methods to accept List<string> instead of domain entities
- Fix syntax errors and missing closing braces
- Fix OSVersion namespace conflicts with System.Environment

Significant progress: 49% error reduction while maintaining hexagonal architecture principles

+++ src/Nexo.Core.Application/Services/AI/Models/ModelManagementService.cs
@@ -280,11 +280,11 @@ namespace Nexo.Core.Application.Services.AI.Models
         {
             return platform switch
             {
-                PlatformType.Web => 650 * 1024 * 1024, // 650MB
-                PlatformType.Desktop => 4 * 1024 * 1024 * 1024, // 4GB
-                PlatformType.Mobile => 1.5 * 1024 * 1024 * 1024, // 1.5GB
-                PlatformType.Console => 2 * 1024 * 1024 * 1024, // 2GB
-                _ => 1 * 1024 * 1024 * 1024 // 1GB default
+                PlatformType.Web => 650L * 1024 * 1024, // 650MB
+                PlatformType.Desktop => 4L * 1024 * 1024 * 1024, // 4GB
+                PlatformType.Mobile => (long)(1.5 * 1024 * 1024 * 1024), // 1.5GB
+                PlatformType.Console => 2L * 1024 * 1024 * 1024, // 2GB
+                _ => 1L * 1024 * 1024 * 1024 // 1GB default
             };
         }
